bump nixinfo version, make appropriate changes

+++ src/output.rs
@@ -9,7 +9,7 @@ fn the_temp(matches: &ArgMatches) -> String {
     if unit == "C" {
         nixinfo::temp().unwrap() + "*C"
     } else if unit == "F" {
-        let pre = nixinfo::temp().unwrap().parse::<i64>().unwrap() * 9 / 5 + 32;
+        let pre = nixinfo::temp().unwrap().parse::<f64>().unwrap() * 9.0 / 5.0 + 32.0;
         pre.to_string() + "*F"
     } else {
         format!("N/A ({} is not a supported unit)", unit)
